+++ src/components/organisms/ValidationSettings/ValidationOverview/ValidationOverview.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import {useValidationSelector} from '@redux/validation/validation.selectors';
+import {useAppSelector} from '@redux/hooks';
+import {isUsingCloudPolicySelector, useValidationSelector} from '@redux/validation/validation.selectors';
 
 import {CRD_SCHEMA_INTEGRATION} from '@shared/models/validationPlugins';
 
@@ -8,14 +9,18 @@ import {VALIDATION_CONFIGURATION_COMPONENTS} from './ConfigurationComponents';
 import CustomValidationCard from './CustomValidationCard';
 import ValidationCard from './ValidationCard';
 import {ValidationCardPlugins} from './ValidationCardPlugins';
+import {ValidationCardPolicy} from './ValidationCardPolicy';
 import ValidationCardUpNext from './ValidationCardUpNext';
 import * as S from './ValidationOverview.styled';
 
 const ValidationOverview: React.FC = () => {
   const plugins = useValidationSelector(s => Object.values(s.metadata ?? {}));
+  const isUsingCloudPolicy = useAppSelector(isUsingCloudPolicySelector);
 
   return (
     <S.ValidationOverviewContainer>
+      {isUsingCloudPolicy && <ValidationCardPolicy />}
+
       {plugins.map(plugin => (
         <ValidationCard
           key={plugin.id}