+++ src/shared/models/rootState.ts
@@ -1,5 +1,6 @@
 import {AlertState} from './alert';
 import {AppState} from './appState';
+import {CloudState} from './cloud';
 import {ClusterState} from './clusterState';
 import {CompareState} from './compare';
 import {AppConfig} from './config';
@@ -28,6 +29,7 @@ type RootState = {
   validation: ValidationState;
   dashboard: DashboardState;
   cluster: ClusterState;
+  cloud: CloudState;
 };
 
 type ElectronMenuDataType = {